chore: move useNotebookLauncher hook logic into a module
