+++ webui/react/src/components/NavigationTabbar.tsx
@@ -4,7 +4,7 @@ import { useLocation } from 'react-router-dom';
 import Auth from 'contexts/Auth';
 import ClusterOverview from 'contexts/ClusterOverview';
 import UI from 'contexts/UI';
-import useNotebookLauncher from 'hooks/useNotebookLauncher';
+import { launchNotebook } from 'utils/task';
 
 import ActionSheet from './ActionSheet';
 import Icon from './Icon';
@@ -39,7 +39,6 @@ const NavigationTabbar: React.FC = () => {
   const { isAuthenticated } = Auth.useStateContext();
   const ui = UI.useStateContext();
   const overview = ClusterOverview.useStateContext();
-  const notebookLauncher = useNotebookLauncher();
   const [ isShowingOverflow, setIsShowingOverflow ] = useState(false);
 
   const cluster = overview.allocation === 0 ? undefined : `${overview.allocation}%`;
@@ -48,9 +47,9 @@ const NavigationTabbar: React.FC = () => {
   const handleOverflowOpen = useCallback(() => setIsShowingOverflow(true), []);
   const handleActionSheetCancel = useCallback(() => setIsShowingOverflow(false), []);
   const handleLaunchNotebook = useCallback((cpuOnly = false) => {
-    cpuOnly ? notebookLauncher.launchCpuOnlyNotebook() : notebookLauncher.launchNotebook();
+    launchNotebook(cpuOnly ? 0 : 1);
     setIsShowingOverflow(false);
-  }, [ notebookLauncher ]);
+  }, []);
 
   if (!showNavigation) return null;
